Cleaning out globals with constants.vals.
